% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{pL_colors}
\alias{pL_colors}
\title{Function to extract pL colors as hex codes}
\format{
An object of class \code{character} of length 7.
}
\usage{
pL_colors
}
\arguments{
\item{...}{Character names of pL_cols}
}
\description{
Function to extract pL colors as hex codes
}
\keyword{datasets}
