% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pL_cols.R
\name{pL_cols}
\alias{pL_cols}
\title{Function to extract pL colors as hex codes}
\usage{
pL_cols(...)
}
\arguments{
\item{...}{Character names of pL_cols}
}
\description{
Function to extract pL colors as hex codes
}
