% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pL_pal.R
\name{pL_pal}
\alias{pL_pal}
\title{Return function to interpolate a pL color palette}
\usage{
pL_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in pL_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a pL color palette
}
