% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_geom_aes_defaults.R
\name{replace_geom_aes_defaults}
\alias{replace_geom_aes_defaults}
\title{Replace default geom colors}
\usage{
replace_geom_aes_defaults(name, old_aes, new_aes)
}
\arguments{
\item{name}{Aesthetic for re-color}

\item{old_aes}{Aesthetic to replace}

\item{new_aes}{replacement Aesthetic to use}
}
\value{
updated ggplot geom colors
}
\description{
Replace default geom colors
}
\examples{
replace_geom_aes_defaults("colour", "black", "#6C6D6C")
}
