% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pL.R
\name{theme_pL}
\alias{theme_pL}
\title{Paegel lab theme for ggplot graphics.}
\usage{
theme_pL()
}
\value{
formatted plot
}
\description{
Formats ggplot graphics in a consistent style for use in notebooks and publications.
}
\examples{
data <- tibble(min = 0:10,  Y = sin(min))
data \%>\% ggplot() + aes(min, Y) + geom_line() + theme_pL()
}
