% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pL.R
\name{theme_pL}
\alias{theme_pL}
\title{Paegel lab theme for ggplot2 graphics.}
\usage{
theme_pL()
}
\value{
formatted plot
}
\description{
Formats ggplot2 graphics in a consistent style for use in notebooks and publications.
}
